Cleaned up /nv command
